Refactor OverlayTrigger imports and create a custom OverlayTrigger component

+++ app/Exceptions/UnableToAllocateStockException.php
@@ -4,9 +4,12 @@ namespace App\Exceptions;
 
 use Exception;
 
+/**
+ * @phpstan-consistent-constructor
+ */
 class UnableToAllocateStockException extends Exception
 {
-    public static function unableToAllocate(string $productName, int $metersRequired, int $metersAllocated): static
+    public static function unableToAllocate(string $productName, float $metersRequired, float $metersAllocated): static
     {
         return new static("Unable to allocate stock for product {$productName} with quantity {$metersRequired} and available stock {$metersAllocated}");
     }
